+++ testutil.go
@@ -4,7 +4,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"math/rand"
+	"runtime"
 	"sort"
+	"strconv"
+	"strings"
 
 	"github.com/sunmxt/sladder/util"
 )
@@ -153,4 +156,15 @@ func (v *TestNamesInKeyIDValidator) Txn(kv KeyValue) (KVTransaction, error) {
 	t.origin, t.changed = kv.Value, false
 	sort.Strings(t.tag.Names)
 	return t, nil
+}
+
+func goid() int {
+	var buf [64]byte
+	n := runtime.Stack(buf[:], false)
+	idField := strings.Fields(strings.TrimPrefix(string(buf[:n]), "goroutine "))[0]
+	id, err := strconv.Atoi(idField)
+	if err != nil {
+		panic(fmt.Sprintf("cannot get goroutine id: %v", err))
+	}
+	return id
 }
